Update: Item triggering system
Reworked the system for using Reaper items to trigger fmod events. Tracking Item's unique GUID identifiers and associating them with unique FMOD event instances so that event instances can have their parameters updated in the middle of their life cycle.

Also updated the AddItemWithLastFMODEvent() to name the inserted item with the event path rather then insert the path as a note in the empty item.