Feature: Parameters can be updated via items with GUID in notes.
Expanded CheckItems to also check for items with names starting with "param:" if found the rest of the name is parsed for the name of the FMOD event parameter and the value (Name=Value). An example of full item name "param:Note=1". When such an item is found its Item Notes are check for a GUID to reference an existing FMOD event to target the parameter updates with.